requirements.txt
scikit-learn
sqlalchemy
psycopg2
pandas
ipykernel
numpy
imbalanced-learn
catboost
xgboost
python-dotenv
requests
pyyaml